Time: 246 ms (95.68%), Space: 99.1 MB (20.93%) - LeetHub

+++ 1838-frequency-of-the-most-frequent-element/1838-frequency-of-the-most-frequent-element.cpp
@@ -4,7 +4,7 @@ public:
         int n = nums.size();
         sort(nums.begin(),nums.end());
         int l=0,r=0;
-        long long total=0;
+        long total=0;
         int ans=0;
         while(r<n){
             total+=nums[r];
